+++ changelog.md
@@ -21,6 +21,7 @@
 * Thorns Enchantment no longer damages items
 * Arrows bypass invinc frames now only works for player's fired arrows
 * Firestarter now takes 3 seconds instead of 4 to light up fire
+* Falling trees now fall better
 
 ## 0.12.0
 * Added Fletching Table